remind_me/requirements.txt
Django==1.8.2
django-cors-headers==1.1.0
django-extensions==1.5.7
django-filter==0.11.0
djangorestframework==3.2.3
Jinja2==2.8
requests==2.7.0
model-mommy==1.2.6
django-rest-auth==0.7.0
pytz==2015.4
arrow==0.7.0